extract http request logic from Fulfillment module
