+++ dhl_fulfillment.rb
@@ -21,4 +21,5 @@ require_relative 'lib/urls/sandbox'
 require_relative 'lib/urls/production'
 require_relative 'lib/exception_utils'
 require_relative 'lib/token_store'
+require_relative 'lib/api_caller'
 require_relative 'lib/dhl_fulfillment'
