+++ lib/dhl_fulfillment.rb
@@ -3,19 +3,15 @@
 module DHL
   # :reek:Attribute
   # :reek:FeatureEnvy
-  # rubocop:disable ModuleLength
   # Connector for the DHL Fulfillment API (https://api-qa.dhlecommerce.com/fulfillment)
   module Fulfillment
-    API_TIMEOUT = 10
     ALREADY_IN_SYSTEM = 'YFC0001'
     INVALID_VALUES_FOR_FIELDS = '919'
 
     class << self
-      include Retry
-
       attr_reader :urls
       attr_accessor :client_id, :client_secret, :account_number
-      attr_writer :token_store
+      attr_writer :api_caller
 
       def configure
         yield self
@@ -30,7 +26,7 @@ module DHL
       end
 
       def create_sales_order(options)
-        call_api method: :post, url: @urls.order_create, body: options.to_json do |response|
+        api.call method: :post, url: @urls.order_create, body: options.to_json do |response|
           raise InvalidValuesFoundForFields if invalid_fields_error?(response)
           raise APIException, "Can't create sales order", response.body unless response.code == 202
           true
@@ -39,7 +35,7 @@ module DHL
 
       def sales_order_acknowledgement(order_number, submission_ud)
         url = order_acknowledgement_url(order_number, submission_ud)
-        call_api method: :get, url: url do |res|
+        api.call method: :get, url: url do |res|
           body = res.body
           raise AlreadyInSystem, order_number, body if already_in_system?(res)
           raise AcknowledgementError, body if acknowledgement_errors?(res)
@@ -49,52 +45,25 @@ module DHL
       end
 
       def sales_order_status(order_number)
-        call_api method: :get, url: order_status_url(order_number) do |res|
+        api.call method: :get, url: order_status_url(order_number) do |res|
           raise APIException, "Can't check sales order status", res.body unless res.code == 200
           true
         end
       end
 
       def shipment_details(order_number)
-        call_api method: :get, url: shipment_details_url(order_number) do |res|
+        api.call method: :get, url: shipment_details_url(order_number) do |res|
           raise APIException, "Can't access shipment details", res.body unless res.code == 200
           true
         end
       end
 
-      def token_store
-        @token_store ||= TokenStore.new(@client_id, @client_secret, @urls)
+      def api
+        token_store = TokenStore.new(@client_id, @client_secret, @urls)
+        @api_caller ||= APICaller.new(token_store)
       end
 
       protected
-
-      def call_api(method:, url:, body: nil, &block)
-        attempt(2).times do
-          begin
-            try_call_api(method, url, body, &block)
-          rescue RestClient::Unauthorized
-            @token_store.clear
-            next_try!
-          end
-        end
-      rescue Retry::OutOfAttempts
-        ExceptionUtils.raise_unauthorized
-      end
-
-      def try_call_api(method, url, body)
-        ExceptionUtils.handle_error_rethrow do
-          response = execute_api_request(method, url, body)
-          yield(response) if block_given?
-        end
-      end
-
-      def execute_api_request(method, url, body)
-        RestClient::Request.execute method: method,
-                                    url: url,
-                                    body: body,
-                                    headers: request_headers,
-                                    timeout: API_TIMEOUT
-      end
 
       def order_acknowledgement_url(order_number, submission_id)
         "#{@urls.order_acknowledgement}/#{@account_number}/#{order_number}/#{submission_id}"
@@ -119,14 +88,6 @@ module DHL
       def api_errors(response)
         payload = JSON.parse(response.body)
         payload.dig('CreationAcknowledge', 'Order', 'OrderSubmission', 'Error') || []
-      end
-
-      def request_headers
-        {
-            'Authorization' => "Bearer #{token_store.api_token}",
-            'Accept'        => 'application/json',
-            'Content-Type'  => 'application/json'
-        }
       end
 
       def invalid_fields_error?(response)
